Moved interfaces to another package
