+++ src/main/java/com/plummy/cyhunters/Assets/GamePlayer.java
@@ -1,6 +1,7 @@
 package com.plummy.cyhunters.Assets;
 
 import com.plummy.cyhunters.Assets.Enums.PlayerState;
+import com.plummy.cyhunters.Assets.Interfaces.IGamePlayer;
 import org.bukkit.entity.Player;
 
 public class GamePlayer implements IGamePlayer {
