--- src/main/java/com/plummy/cyhunters/Assets/IGameBoard.java
@@ -1,4 +1,0 @@
-package com.plummy.cyhunters.Assets;
-
-public interface IGameBoard {
-}
