+++ src/main/java/com/plummy/cyhunters/Assets/Interfaces/IGameBoard.java
@@ -1,0 +1,4 @@
+package com.plummy.cyhunters.Assets.Interfaces;
+
+public interface IGameBoard {
+}
